+++ src/design/plone/contenttypes/behaviors/evento.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from collective import dexteritytextindexer
+from plone.app.dexterity import textindexer
 from design.plone.contenttypes import _
 from collective.volto.blocksfield.field import BlocksField
 from plone.app.z3cform.widget import RelatedItemsFieldWidget
@@ -230,7 +230,7 @@ class IEvento(model.Schema):
         fields=["patrocinato_da"],
     )
 
-    dexteritytextindexer.searchable("descrizione_estesa")
+    textindexer.searchable("descrizione_estesa")
 
 
 @implementer(IEvento)